+++ framework/model-test/src/main/java/is/codion/framework/model/test/AbstractEntityModelTest.java
@@ -307,12 +307,12 @@ public abstract class AbstractEntityModelTest<Model extends DefaultEntityModel<M
 
 		Model model = createDepartmentModelWithoutDetailModel();
 		model.editModel().insert(asList(dept, emp));
-		assertTrue(model.tableModel().containsItem(dept));
-		assertFalse(model.tableModel().containsItem(emp));
+		assertTrue(model.tableModel().items().contains(dept));
+		assertFalse(model.tableModel().items().contains(emp));
 
 		model.editModel().delete(asList(dept, emp));
 
-		assertFalse(model.tableModel().containsItem(dept));
+		assertFalse(model.tableModel().items().contains(dept));
 	}
 
 	protected final EntityConnectionProvider connectionProvider() {