+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/combobox/DefaultItemComboBoxBuilder.java
@@ -207,7 +207,7 @@ final class DefaultItemComboBoxBuilder<T> extends AbstractComponentBuilder<T, JC
 				comboBoxModel = itemComboBoxModel(modelItems);
 			}
 		}
-		if (nullable && comboBoxModel.containsItem(nullItem)) {
+		if (nullable && comboBoxModel.items().contains(nullItem)) {
 			comboBoxModel.setSelectedItem(null);
 		}