+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityEditModelTest.java
@@ -130,7 +130,7 @@ public class SwingEntityEditModelTest {
 		comboBoxModel.refresh();
 		assertEquals(4, comboBoxModel.getSize());
 		for (EnumType enumType : EnumType.values()) {
-			assertTrue(comboBoxModel.containsItem(enumType));
+			assertTrue(comboBoxModel.items().contains(enumType));
 		}
 	}
 }